# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Investintell Light API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/builder/optimize": {
      "post": {
        "tags": [
          "builder"
        ],
        "summary": "Optimize",
        "description": "Optimize weights over a mixed fund/equity universe.\n\nDefault objective is ``min_cvar`` (Rockafellar\u2013Uryasev, \u03b1=0.95) on raw\nhistorical scenarios. With Black-Litterman ``views``, scenarios are\nre-centered on the posterior \u03bc_BL and floored at the equilibrium return;\n``bl_utility`` selects the explicit max-utility objective instead.\nAll fractional fields are decimal fractions (0.05 = 5%).",
        "operationId": "optimize_builder_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/builder/save": {
      "post": {
        "tags": [
          "builder"
        ],
        "summary": "Save",
        "description": "Persist a builder proposal as a real portfolio (F8.5 + F8.6b).\n\nEach weight is sized at the asset's REFERENCE price (equities: latest\nadj_close; funds: latest NAV) against ``notional_usd`` \u2014 unless it\ncarries a ``fill_price``, which sizes the position and (with the\ncommission) defines the real cost basis (basis='executed'). Fund weights\nmay select a share class via ``class_ticker`` (same instrument; priced\nwith the series NAV as a proxy). The portfolio is tagged\norigin='builder'. Domain failures \u2014 asset without a price, fund without\na ticker, invalid class, duplicate portfolio name, weight too small for\nthe notional \u2014 are 422 verbatim.",
        "operationId": "save_builder_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AbsoluteViewIn": {
        "properties": {
          "type": {
            "type": "string",
            "const": "absolute",
            "title": "Type"
          },
          "asset": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FundRefIn"
              },
              {
                "$ref": "#/components/schemas/EquityRefIn"
              }
            ],
            "title": "Asset",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "equity": "#/components/schemas/EquityRefIn",
                "fund": "#/components/schemas/FundRefIn"
              }
            }
          },
          "q": {
            "type": "number",
            "title": "Q"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Confidence",
            "default": 0.5
          }
        },
        "type": "object",
        "required": [
          "type",
          "asset",
          "q"
        ],
        "title": "AbsoluteViewIn",
        "description": "'Asset returns q per year' (e.g. q=0.12 \u2192 12% a.a.)."
      },
...
      "BLParamsIn": {
        "properties": {
          "delta": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Delta",
            "default": 2.5
          },
          "tau": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Tau",
            "default": 0.05
          }
        },
        "type": "object",
        "title": "BLParamsIn"
      },
...
      "ConstraintsIn": {
        "properties": {
          "cap": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cap",
            "default": 0.25
          },
          "min_weight": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Weight"
          }
        },
        "type": "object",
        "title": "ConstraintsIn",
        "description": "Long-only and sum(w)=1 are always enforced; these are the knobs."
      },
...
      "DiagnosticsOut": {
        "properties": {
          "n_obs": {
            "type": "integer",
            "title": "N Obs"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "mu_equilibrium": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mu Equilibrium"
          },
          "mu_posterior": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mu Posterior"
          }
        },
        "type": "object",
        "required": [
          "n_obs",
          "status"
        ],
        "title": "DiagnosticsOut"
      },
...
      "EquityRefIn": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "equity",
            "title": "Kind"
          },
          "ticker": {
            "type": "string",
            "maxLength": 12,
            "minLength": 1,
            "title": "Ticker"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "ticker"
        ],
        "title": "EquityRefIn"
      },
      "ExpectedOut": {
        "properties": {
          "vol_ann": {
            "type": "number",
            "title": "Vol Ann"
          },
          "cvar_95_in_sample": {
            "type": "number",
            "title": "Cvar 95 In Sample"
          },
          "return_ann_bl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Ann Bl"
          }
        },
        "type": "object",
        "required": [
          "vol_ann",
          "cvar_95_in_sample",
          "return_ann_bl"
        ],
        "title": "ExpectedOut"
      },
...
      "FundRefIn": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "fund",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "title": "FundRefIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizeRequest": {
        "properties": {
          "assets": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/FundRefIn"
                },
                {
                  "$ref": "#/components/schemas/EquityRefIn"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "equity": "#/components/schemas/EquityRefIn",
                  "fund": "#/components/schemas/FundRefIn"
                }
              }
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 2,
            "title": "Assets"
          },
          "objective": {
            "type": "string",
            "enum": [
              "equal_weight",
              "min_vol",
              "erc",
              "max_diversification",
              "min_cvar",
              "bl_utility"
            ],
            "title": "Objective",
            "default": "min_cvar"
          },
          "constraints": {
            "$ref": "#/components/schemas/ConstraintsIn",
            "default": {
              "cap": 0.25
            }
          },
          "window_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 30.0,
            "title": "Window Days",
            "default": 730
          },
          "views": {
            "anyOf": [
              {
                "items": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AbsoluteViewIn"
                    },
                    {
                      "$ref": "#/components/schemas/RelativeViewIn"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "absolute": "#/components/schemas/AbsoluteViewIn",
                      "relative": "#/components/schemas/RelativeViewIn"
                    }
                  }
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Views"
          },
          "bl": {
            "$ref": "#/components/schemas/BLParamsIn",
            "default": {
              "delta": 2.5,
              "tau": 0.05
            }
          }
        },
        "type": "object",
        "required": [
          "assets"
        ],
        "title": "OptimizeRequest"
      },
      "OptimizeResponse": {
        "properties": {
          "weights": {
            "items": {
              "$ref": "#/components/schemas/WeightOut"
            },
            "type": "array",
            "title": "Weights"
          },
          "expected": {
            "$ref": "#/components/schemas/ExpectedOut"
          },
          "diagnostics": {
            "$ref": "#/components/schemas/DiagnosticsOut"
          }
        },
        "type": "object",
        "required": [
          "weights",
          "expected",
          "diagnostics"
        ],
        "title": "OptimizeResponse"
      },
...
      "RelativeViewIn": {
        "properties": {
          "type": {
            "type": "string",
            "const": "relative",
            "title": "Type"
          },
          "long": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FundRefIn"
              },
              {
                "$ref": "#/components/schemas/EquityRefIn"
              }
            ],
            "title": "Long",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "equity": "#/components/schemas/EquityRefIn",
                "fund": "#/components/schemas/FundRefIn"
              }
            }
          },
          "short": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FundRefIn"
              },
              {
                "$ref": "#/components/schemas/EquityRefIn"
              }
            ],
            "title": "Short",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "equity": "#/components/schemas/EquityRefIn",
                "fund": "#/components/schemas/FundRefIn"
              }
            }
          },
          "q": {
            "type": "number",
            "title": "Q"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Confidence",
            "default": 0.5
          }
        },
        "type": "object",
        "required": [
          "type",
          "long",
          "short",
          "q"
        ],
        "title": "RelativeViewIn",
        "description": "'`long` outperforms `short` by q per year'."
      },
...
      "SaveRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Portfolio name; same rules as POST /portfolios."
          },
          "notional_usd": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Notional Usd",
            "default": 1000000
          },
          "weights": {
            "items": {
              "$ref": "#/components/schemas/SaveWeightIn"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Weights"
          }
        },
        "type": "object",
        "required": [
          "name",
          "weights"
        ],
        "title": "SaveRequest",
        "description": "Body for POST /builder/save \u2014 persist a proposal as a real portfolio.\n\n``quantity = weight * notional_usd / price`` per position (rounded to\n4 decimals), where price is the fill price when given, else the\nreference spot/NAV. The stored ``acq_price`` is the reference price for\nbasis='reference', or ``(fill_price*qty + commission)/qty`` (6 decimals)\nfor basis='executed'."
      },
      "SaveResponse": {
        "properties": {
          "portfolio_id": {
            "type": "integer",
            "title": "Portfolio Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "notional_usd": {
            "type": "number",
            "title": "Notional Usd"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/SavedPositionOut"
            },
            "type": "array",
            "title": "Positions"
          },
          "pricing_note": {
            "type": "string",
            "title": "Pricing Note",
            "default": "Reference prices (spot/NAV) are for analysis; executed fills with commissions define real cost basis. Fund class NAV is proxied by the series NAV."
          }
        },
        "type": "object",
        "required": [
          "portfolio_id",
          "name",
          "notional_usd",
          "positions"
        ],
        "title": "SaveResponse"
      },
      "SaveWeightIn": {
        "properties": {
          "asset": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FundRefIn"
              },
              {
                "$ref": "#/components/schemas/EquityRefIn"
              }
            ],
            "title": "Asset",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "equity": "#/components/schemas/EquityRefIn",
                "fund": "#/components/schemas/FundRefIn"
              }
            }
          },
          "weight": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Weight"
          },
          "fill_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Price",
            "description": "Actual execution price per share/unit; presence flips the position to basis='executed'."
          },
          "commission": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission",
            "description": "Total commission paid on the fill (>= 0); requires fill_price."
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date",
            "description": "Execution date of the fill; requires fill_price."
          },
          "class_ticker": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 12,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Ticker",
            "description": "Fund share-class ticker (fund assets only); must belong to the same fund instrument. Priced with the series NAV as a proxy."
          }
        },
        "type": "object",
        "required": [
          "asset",
          "weight"
        ],
        "title": "SaveWeightIn",
        "description": "One proposed weight to persist. Zero/near-zero weights should be\nfiltered out by the caller \u2014 a weight that rounds to quantity 0 is a 422.\n\nF8.6b execution fields (all optional, retro-compatible):\n- without ``fill_price`` the position is saved at the REFERENCE price\n  (spot/NAV) with basis='reference';\n- with ``fill_price`` the position is EXECUTED: quantity is sized at the\n  fill, and the cost basis includes the commission;\n- ``class_ticker`` (funds only) saves the position under a share-class\n  ticker of the SAME fund instead of the representative one."
      },
      "SavedPositionOut": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "basis": {
            "type": "string",
            "enum": [
              "reference",
              "executed"
            ],
            "title": "Basis"
          },
          "cost_basis": {
            "type": "number",
            "title": "Cost Basis"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "price",
          "basis",
          "cost_basis"
        ],
        "title": "SavedPositionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeightOut": {
        "properties": {
          "asset": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FundRefIn"
              },
              {
                "$ref": "#/components/schemas/EquityRefIn"
              }
            ],
            "title": "Asset",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "equity": "#/components/schemas/EquityRefIn",
                "fund": "#/components/schemas/FundRefIn"
              }
            }
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "asset",
          "weight"
        ],
        "title": "WeightOut"
      }
    }
  }
}